Polishing
- C++ casts
- better README
- base
  * Mat33 and Mat44 are structs now
  * Quat extends Vec3
- matrix
  * destroy events
  * Physics: better hit event handling, splash event
- client
  * Sky: occlusion, optimized rendering
  * smoother MD2 animation
  * simplified BasicSound, one hit, one splash sound

+++ src/base/Math.cpp
@@ -19,7 +19,7 @@ namespace oz
 
   Math::Math()
   {
-    srand( (uint) time( null ) );
+    srand( time( null ) );
   }
 
   void Math::seed( uint seed )
